renamed projectslist to dashboard
