+++ client/src/pages/Dashboard/Dashboard.js
@@ -1,17 +1,37 @@
 import React from "react";
-import GoogleAuth from "../../components/GoogleAuth/GoogleAuth";
 import { Link } from "react-router-dom";
 import "./Dashboard.scss";
 
 const Dashboard = () => {
+  const projects = [1, 2, 3, 4, 5];
   return (
-    <div className="dashboard">
-      <p>Dashboard Page</p>
-      <Link to="/">Home Page</Link>
-      <Link to="login">Login Page</Link>
-      <Link to="dashboard">Dashboard Page</Link>
-      <Link to="signup">Register Page</Link>
-      <GoogleAuth className="login-page" />
+    <div className="wrapper">
+      <section className="header">
+        <p>Crowdfund</p>
+        <div className="header__right">
+          <p>About</p>
+          <p>Discover</p>
+          <p>Get Started</p>
+        </div>
+      </section>
+      <section className="main">
+        <div className="main__left">
+          <div className="main__left--info">
+            <p>Settings</p>
+            <p>Fundraiser</p>
+            <p>Donations</p>
+            <p>Payment Infos</p>
+          </div>
+        </div>
+        <div className="main__right">
+          {projects.map((el) => (
+            <div className="main__right--card">
+              <div>Image: </div>
+              <div>Text: </div>
+            </div>
+          ))}
+        </div>
+      </section>
     </div>
   );
 };
